fixup: move redirect methods form Url to Redirect
